+++ src/borrow_pcg/mod.rs
@@ -15,5 +15,4 @@ pub mod has_pcs_elem;
 pub mod latest;
 pub mod path_condition;
 pub mod region_projection;
-pub mod region_projection_member;
 pub mod unblock_graph;
